requirements.txt
asgiref==3.4.1
Django==3.2.5
django-filter==2.4.0
idna==3.2
pytz==2021.1
sqlparse==0.4.1

djangorestframework~=3.12.4
django-debug-toolbar~=3.2.1